linkedin/requirements.txt
linkedin-api-client==0.3.0